305.2. Connecting Redux Form(Section 23: Handling Forms with Redux Form)
So we are importing a named import right here and then we're going to essentially rename it on the fly

on the exact same line of code.

So we will not have a variable inside this file called reducer.

Instead it's going to be imported into this file and given a name of form reducer.

Again this is done just to make sure that it's kind of clear what reducer we are talking about.

At any given point because you might have a lot of different reducers inside this file.

And if you just see one side of your called reducer you would probably say what in the world is this

reducer.

So we're just trying to give it a more descriptive name that's all.

So if we use this syntax right here we no longer have a variable called reducer and we instead have

to write out form reducer like XA OK.

So let's now save this file.

We're going to flip back over to our application and just refresh and make sure that we don't see any

error message.

You will see an error message.

If you assign that reducer to a key other than form inside of your combined reducers call.

Now if you open up your redux dev tools.

Now here's my redux dev tools over here in on making sure that I'm looking at my react up.

You'll notice that if you select state right here and look at tree we have a new key inside of our state

tree or redux store that new key is forms.

That's the reducer that we just set up over time as we start to put foreigns together inside of application.

We're going to see that this piece of state right year has a bunch of data related to the forms that

we have inside of our app.

So as we start to wire together this create form that we're trying to build this one right here.

We're going to see that as we start to type text in here.

All that text and whatnot is going to automatically show up inside of this form object right there inside

of a redux store.

Remember the entire idea here is that with redux form we are still using regex all the same redux you

know in understand just a lot.

The steps are being done for us automatically OK now that we have wired up Bre-X form.

We're going to take a quick pause when we come back the next video.

We're going to start working on our create stream form and understand how to work with redux form in

general.

So quit posin I'll see you in just a minute.