306.1. Creating Forms(Section 23: Handling Forms with Redux Form)
In this video we're going to start working on our stream create form.

The first thing we do is make sure that we can see the stream create a component on our screen.

At present we are just at local whoas 3000 which by default always shows the stream list.

So I'm going to navigate over to local host 3000 slash streams slash new like so and then I should see

it stream create appear on the screen.

Remember this only works because we had set up our re-act router inside of our app component right here.

So here's the route where we gave a path of stream slash knew.

We said that whenever a user went to that route we wanted to show the component stream create and that's

exactly what's happening right now.

Now we can see this component on the screen.

I'm going to find it inside of my code editor in the components directory.

I'll find the streams folder and then I'll open up the stream create file.

All right.

So we're going to be doing a lot of code or writing a lot of code inside of here to show our form and

make use of redux form as well.

Now the first thing I want to do inside of here is refactor it.

This functional component overt to a class based component.

The reason that we're going to make a class based component is that we're going to eventually want to

have a bunch of helper methods so we can kind of better organize our code inside of here.

All right.

So I'm going to delete that first line and say stream create extends re-act component and then open

up my curly brace up there.

I'll put in my render method and it's going to just wrap the return statement and then I'm going to

delete the semi-colon at the end of the class declaration.

So if I now save this I should be able to flip back over and still see a stream create appear on the

screen.

All right.

Next up we're going to import in redux form at the top.