306.2. Creating Forms(Section 23: Handling Forms with Redux Form)
I'm going to import two helpers or two properties from redux form itself.

The first one is fields with a capital F and then the second is redux ormer with a lowercase are from

redux Dasch form now a set.

These things have different capitalization.

Field has a capital F. redux form is a lowercase are the reason for this is that field is supposed to

be a re-act component.

So this is going to be a component that we are going to eventually show on the screen.

On the other hand redux form right here is a function.

This function is going to essentially have the exact same functionality as the connect function that

used from the re-act redux libron the redux form function is essentially what's going to make sure that

we can call some action creator and get some form data into our component.

And remember that all happens automatically it's the first thing we're going to do is take this redux

form function and hook it up to our component down here at the very bottom again redux form very similar

nature to connect so similar in nature that it even has an identical syntax when we make use of it.

So in front of string creates I'm going to write out redux form.

I'm going to put in one set of parentheses and then put a second set of parentheses around stream create.

Remember the syntax here is that redux form is going to return a function and we immediately call that

function with stream create.

Now we're going to pass redux form right here an object for some configuration unlike the connect function

which takes separate arguments into its redux form instead receives a single object and we put a bunch

of configuration into that object.

At present there is only one piece of configuration we have to put in here.

I'm going to say form and then I'm going to provide a name for this form as a stream string.

The name of this form is going to be generally whatever the purpose of the form is.

You'll see what the purpose of this name is in just a minute.

So for right now I'll just put in and name something like.

Stream creates it can be lowercase capitalized whatever you want it to be.

It really doesn't matter.

OK so now we have hooked up redux form.

This component is now going to be passed.

A ton of props a ton of additional props that it was not being passed before just to get a better handle

on what's going on.

Let's add in a console log statement inside the render method and I'm going to count the log.

Stop props all then flip back over

if you see any errors or warnings like this again you're going to see those errors or warnings every

now and then when you were making use of the Google API stuff.

If you see those errors and warnings totally fine you can ignore them altogether.

All right so I get to refresh the page and then I should see this council aigrette year of my prop's

object.

So this is our prop's object that is going into that form.

Notice how there are a tremendous number of properties on that process object.

Now before we probably only had a handful now we have a tremendous number of props that are showing

up.

Thanks to that redux form helper all the props you see right here are part of the automatic system that's

going to be used to essentially get form values into a element and then get changes back out and update

that redux form reducer.

So it's essentially up to you and me to kind of pick through this list of props right here and find

the different props that are relevant to us and make use of them to actually build out our form.

All right.

So I'm going to flip back over.