306.3. Creating Forms(Section 23: Handling Forms with Redux Form)
I'm going to remove that council log and then I'm going to turn this return statement right here into

a multi-line GSX expression.

So I'll put in a set Apprentice's like so I'm going to return a form element and inside if you're going

to put together something called a field.

Now this field component right here has been imported from redux form at the top.

So this is a component that was not created by us.

We make use of this field component anytime that we want to show a.

As you might guess a field to the user a field is essentially some type of input.

It can be a checkbox or a radio button or a text input or a dropdown just about anything that is going

to solicit input from a user.

Whenever we place a field tag we're always going to provide it some number of prompts.

There's one prof that is always required and that is the name Propp the name Propp is not going to be

like the title or the label or the field the name is like the name of the property that this field is

going to manage the form that you and I are trying to put together.

It's going to have a title and a description.

So for our first input right here the first input is going to be managing the title of our stream.

So I'm going to give this first field right here a name of title because I want this field to be responsible

for handling the input from the user to indicate the title of the stream or then put down a second field

so the second field right here is meant to handle this description input right here.

So I'm going to give this one a name of description like so now when you start making use of redux form

there are a couple of error messages that you're probably going to see really frequently.

So I want to show you what those error messages right now.

We're going to save this and then we're going to flip back over to our application and just see what

happens.

So you're going to see this error message right here it says expected a string or a class function but

got undefined.

So what's going on here.

Well essentially when we show this field component the field itself has no idea how to render some type

of input element on the screen.

So a field doesn't know how to show a text input.

It doesn't actually know how to show a checkbox it doesn't know how to show it dropped down or anything

like that.

A field is just a component that's going to be part of the system that's going to automatically handle

all of the forms the put together but it individually is not responsible for actually getting anything

to show up on the screen.

Now let's take a quick pause right here before the section gets too long.

We're going to come back the next section and talk about how we can actually get a text input to appear

when we are making use of a field element as in both these cases right here.

So quit pause and I'll see you in just a minute.