307.1. Automatically Handling Events(Section 23: Handling Forms with Redux Form)
In the last video we started hooking up redux form to our history create component.

We placed these two field components inside of our component but we very quickly saw this error message

inside of our browser thing that says expected a string or class function but got undefined.

So the reason that we are seeing this is that the field component itself doesn't actually know how to

show a text input or a dropdown or a checkbox or any type of form element in order to tell this field

how to show an actual text input on the screen.

We have to assign a prop to it called component component right here is going to be either a real component

or a function for the field to call this component or the function needs to return some element that

is going to actually be shown on the screen.

So this field element in total is really just about hooking up all of this infrastructure all the stuff

that redux form is actually doing for us when we actually want to show some content or some form element

on the screen.

It is up to us to show the actual input element and customize it in some fashion.

And we do so by making use of the component prop right here so inside of my stream create class I'm

going to add in a helper method called render input and then I'm going to pass a reference to that function

to the component right here.

So this Dom render input and I'm going to do the same thing for the second field as well.

Just render input then inside of render input right here.

I'm going to return a div that says I'm an input like so.

Now clearly that is not in input right.

I think you can agree with me there.

Well I want to prove a little point to you here.

Let's say this or we're going to flip back over to our browser.

When I flip back over I see on the screen twice I'm an input.

So this is a very important lesson to learn.

Remember redux form is about automating all these processes making sure that your form data is inside

the read store getting that data back into the component and then kind of helping you getting that data

into the appropriate input elements.

But at the end of the day redux farm doesn't really know a lot about what you're doing.

It doesn't really know if you are showing a text input.

It doesn't know if you're showing a dropdown it doesn't know if you're showing a checkbox it doesn't

know if you're not showing any input element at all.

Redux farm is just kind of going to assume that you are doing the correct thing.

So in this case for us to do the correct thing we are really supposed to return an input element from

whatever function we passed to the component Propp right here.