307.2. Automatically Handling Events(Section 23: Handling Forms with Redux Form)
So to do things the right way instead of returning a div we're going to return an input element like

so.

So I'm going to save this and I'll flip back over to my browser and I can see that now had those two

input elements.

Now we're not done quite yet.

Remember any time that we show an input element with react in any capacity be it with react or redux

a redux form we always are supposed to make sure that we show that input element in assign its value

property and also give it in on change callback handler.

So right now we are not doing any of those things.

We are just showing a plane input element and this is not a controlled element as we would prefer it

to be.

So in order to turn this into a controlled element any time that the field tag right here calls render

input it's going to pass in some number of arguments right here.

So I'm going to receive this as a single argument called form props and let's just cancel that out right

now and see what we are dealing with.

It's going to flip back over and we could see that we got to cancel logs right here we get to count

locks because we are showing two separate inputs and that is causing render input to be called two times

so I'm going to expand.

But the first one right here inside this object.

You'll notice that we have keys of inputs and meida the input property right here has our actual field

name of title.

So this object right here is meant to essentially be customizing this first input the second consulate.

The second input right here is a name of description.

So this right here is meant to somehow customize the second input.

Over here you'll notice that inside of this input element to this input property we have a bunch of

very familiar looking callback handlers.

So for example if your is on change and we all set the value prop right there as well.

So the intent here.

Again can't say it enough is that redux form is going to wire up all the stuff for us it is our job

to make this very last connection.

It is our job to take that onchange handler and wire it up to the input element and it is our job to

take the value prop and hook it up to the input element as well.