307.3. Automatically Handling Events(Section 23: Handling Forms with Redux Form)
So essentially you and I are supposed to take this form props take the input property out of it and

hook up the relevant properties to the text input.

So we can say something like say on change we want to call form prop's input on change and then for

the value we would pass in a forum prop's input value like so now when I say this it's going to turn

it into a multi-line expression like so totally equivalent to what you just saw on the screen a moment

ago.

OK.

So let's now flip back over to our application and now see what's happening.

When I flip back over everything looks the same but now really interestingly if you type something in

here you can then open up your redux dev tools and you're going to start to see some really interesting

input over there.

OK so I'm going to expand my tools here and they'll notice that when we typed in characters we got a

ton of these different change action types dispatched to our redux store.

So every one of these different change actions that you see is a different key press.

When I typed into that text input in addition if you select the state selector over here on the top

right hand side and then select the tree selector right here you can see that the form piece of state

right here now has some additional property inside of it before it just had one empty object.

I can now open up this object I'll see that it has a key of string create that string create key right

there is coming from the name of the form that we put down here and then if we open up that object we'll

see that it has a Values property and has the title and the description out of our form.

So this is how redux form it stores data inside of your reducer that you wired up to the combined reducers

call every time that you create a new form read this read Form call right here.

You pass in that name whatever name you put inside there read redux form is going to start to store

all the values for that form on a key with that name inside of the form reducer and then nested somewhere

inside there is going to be all the different values out of your form.

So this is just normal redux nothing different than you have learned already.

This is just redux form storing all this information for you automatically.

That is it all right.

It's not going to condense that down on the right hand side.

So clearly redux form is now kind of handling text input for both these two inputs right here.

And we just did that by forcibly taking all the props or two props I should say out of the form prop's

argument right here and passing them into the input arguments for the onchange and value properties.