307.4. Automatically Handling Events(Section 23: Handling Forms with Redux Form)
When he starts to look at the documentation there is actually going to be a little shortened way of

doing this.

So we do not actually have to specify like oh yeah take the onchange prop and assign it to on change

and take the value prop and assign it to value instead.

We can use a little shortened syntax to do this to do so.

I'm going to remove both those props and I'm going to turn this back into a single line.

GSX expression like so.

It's now inside of.

Going to see a little bit of new syntax that we've never seen around GSX before.

I'm going to put down a set of curly braces without any equals like so in front of it.

So just a set of curly braces and then inside if you're going to say Dot dot dot form prop's dot input

like so.

So again this is a new little bit of syntax.

This essentially is going to take the form prop's input property.

So that was an object with the value prop where the value property and the on change handler in all

that stuff that's going to take all those key value pairs and add them as properties to the input element.

So essentially take that entire input object take all the properties out there and add them as props

to the input element.

So if we now say this again we can flip back over.

I can enter in some text here and again we're going to see the same thing that we saw before.

We now see it we have the form with string create and it has values on it of whatever we just entered

in there.

Now we also see some other properties inside there now that we are correctly passing and all the props.

So any time that we make use of redux form we are always going to use this syntax because there are

some other properties inside that input object besides just the value one and the onchange one that

redux form cares about.

Now as a way of shortening up the syntax even further we can actually structure the input argument out

of the form prop's object.

So I'm going to structure out input.

And now we can reduce this down to just be dot dot dot input like so.

OK so I'm going to save this right here.

We're going to take another quick break.

We'll come back the next video and we're going to continue working on our form.