307.5. Automatically Handling Events(Section 23: Handling Forms with Redux Form)
Now as a way of shortening up the syntax even further we can actually structure the input argument out

of the form prop's object.

So I'm going to structure out input.

And now we can reduce this down to just be dot dot dot input like so.

OK so I'm going to save this right here.

We're going to take another quick break.

We'll come back the next video and we're going to continue working on our form.