309.1. Handling Form Submission(Section 23: Handling Forms with Redux Form)
We've now got our two fields put together.

So we're now going to start to pivot our attention to handling submission of this form.

Remember at some point in time the user is going to want to click a button to actually create a stream

or do whatever it is the form is supposed to do.

We're going to first begin by approaching this problem in the same way that we have handled the form

submission in the past and then we're going to very quickly refactor it to use redux form as well.

So I'm going to put together a little helper method inside my class called on submit and remember anytime

that this thing gets called You'll be called with a event object.

Usually we want to call something like event or prevent default on that object to keep it from attempting

to refresh the browser as the browser tries to submit the form to some back end server that doesn't

actually exist in this case.

So again this is what we usually do but now that we are using redux form things are going to be just

a little bit different.

So as a quick reminder just a little bit ago we did a console log inside of our render method of this

stock drops.

I'm going to test that out again.

I'm going to save this and we're going to take a look at that prop's object really quickly one more

time.

So back inside my browser Here's my prop's object.

And as before we see that there are a tremendous number of methods or excuse me properties added to

that prop's object I'll bite redux form if you look inside if you're going to see that there's one particular

property right here called handle submit.

So this is a function that we are supposed to call any time that the form gets submitted.

This function is provided by redux form.

You'll see what it does.

In just a moment.

It's a lot easier to just kind of show you what this does rather than give you a kind of unclear definition.

So let's just write out some code here and you'll very quickly understand what's going on.

OK.