309.2. Handling Form Submission(Section 23: Handling Forms with Redux Form)
So I'm going to delete the counsel log and then going to find the form element right here and I'm going

to add on the on submit Propp like so.

Now here's where things get interesting to submit.

Usually we would pass an on submit our callback right here but now that we are using redux form things

are going to be just a bit different.

So to on submit.

I'm going to pass this.

Prop stock handles submit and then into that I'm going to pass my this on submit function like so.

So a lot of very similar terminology going on right here.

Let's just very quickly go over that block.

So we have on submit that is the name of the prop that we are passing down to the form if we pass a

function called are osme if we have a function on the on Samit Propp down into the form the function

we called any time that the form gets submitted then inside of here we first have this prop's not handle

Samit.

This is a callback function that is provided to our component by redux form.

We then call that function with all recall back method that we defined inside of our component right

here.

And when we go through this entire process it's going to change exactly how on Samake It's called internally.