309.3. Handling Form Submission(Section 23: Handling Forms with Redux Form)
When ever this form gets submitted now handle Smit is going to automatically receive that event object

and automatically call prevent default for us.

That is something that's really tedious for us to have to do a redux form just takes care of it for

us.

So we don't have to call that anymore.

In fact on Samit it's not even going to be called with an event object at all.

Instead it's going to be called with whatever values existed inside of our two field inputs.

So usually we might refer to this argument right here as something like form props now form prop's does

not mean that this is some data passed down by a parent component.

Maybe a better name this argument here would be something like safe form values.

Let's stick with that right now just make things a little bit more clear.

So let's console log out those form values agile very quickly get a good idea of what's going on here.

Now Leslie I want to do very quickly inside the form I'm going to put in a button with the class name

of UI button primary

and I'll give this thing some text like Samit.

All right let's flip back over and we'll test out really quickly.

So enter in some title and some description.

Then when I click on submit there's my council log.

So I've got the title and the description that were entered into the form.

No longer are we dealing with some event object as our argument right here.

The object is really not useful to us one bit.

The only reason we ever dealt with the event object was so that we could call that present default method.

So now this on Samit callback rather than getting the useless event object is going to be called with

all the values out of our form inside of an object and that's going to make it really easy for us to

take those form values now and save them on some API or validate or whatever it is we want to do with

them.

All right.

So that's pretty much it on how we handle forms submission with redux form.

Again we reference handle Samit and then we pass in whatever callback we want to run after the form

gets submitted.

So in reality handles submit is going to be called.

And then once it processes the form event and all that kind of good stuff our callback then gets invoked

with our actual form values that you and I care about.

All right.

This looks pretty good.

Now one little thing I want to point out here if I refresh the browser and I just click on submit right

away I still get my submission event even though I did not enter a title in a description.

The chances are that you and I do not want people to be able to create a stream that does not have a

title and a description.

So I think that before we allow a user to submit this form we should really be validating these text

inputs.

So let's take a pause right here.

When we come back the next section we're going to learn how we can validate some text inputs with redux

form.