311.1. Displaying Validation Messages(Section 23: Handling Forms with Redux Form)
In the last video we created a Validate function.

Remember this thing is going to be called with all the different values out of our form.

So it's our job to create an Aries object attach any error messages that we might need to show to the

user if they entered in invalid inputs and then we return that is object at the bottom.

Now we need to make sure that this ballot function gets wired up to redux form so that it knows to use

the validate function.

So we're going to take that validate function and we're going to pass it into the redux form helper

down here on a key called validate.

So I'll say validate is the validate function we just created.

Now notice how the key and the value here are identical.